Render ICU Status Report
This YAML code specifies the name of the GitHub Action, when it will run, what type of machine it will run on, and the steps the action will take when it runs. It also specifies that the action will use R version 4.1.0 and install the necessary R packages. The Rscript command renders the icu_status_report.Rmd file to HTML, and the git-auto-commit-action automatically commits the changes to the forked repository.